+++ demo/admin/src/pages/PageContentBlock.tsx
@@ -1,14 +1,14 @@
 import { createBlocksBlock, SpaceBlock, YouTubeVideoBlock } from "@comet/blocks-admin";
 import { AnchorBlock, DamImageBlock, DamVideoBlock } from "@comet/cms-admin";
+import { HeadlineBlock } from "@src/common/blocks/HeadlineBlock";
 import { LinkListBlock } from "@src/common/blocks/LinkListBlock";
 import { RichTextBlock } from "@src/common/blocks/RichTextBlock";
+import { TextImageBlock } from "@src/common/blocks/TextImageBlock";
 import { userGroupAdditionalItemFields } from "@src/userGroups/userGroupAdditionalItemFields";
 import { UserGroupChip } from "@src/userGroups/UserGroupChip";
 import { UserGroupContextMenuItem } from "@src/userGroups/UserGroupContextMenuItem";
 import * as React from "react";
 
-import { HeadlineBlock } from "../common/blocks/HeadlineBlock";
-import { TextImageBlock } from "../common/blocks/TextImageBlock";
 import { ColumnsBlock } from "./blocks/ColumnsBlock";
 import { FullWidthImageBlock } from "./blocks/FullWidthImageBlock";
 import { MediaBlock } from "./blocks/MediaBlock";
@@ -37,13 +37,9 @@ export const PageContentBlock = createBlocksBlock({
         ...userGroupAdditionalItemFields,
     },
     AdditionalItemContextMenuItems: ({ item, onChange, onMenuClose }) => {
-        // TODO fix typing: infer additional fields somehow
-        // @ts-expect-error missing additional field
         return <UserGroupContextMenuItem item={item} onChange={onChange} onMenuClose={onMenuClose} />;
     },
     AdditionalItemContent: ({ item }) => {
-        // TODO fix typing: infer additional fields somehow
-        // @ts-expect-error missing additional field
         return <UserGroupChip item={item} />;
     },
 });