+++ packages/admin/cms-admin/src/pages/pagesPage/createPagesQuery.ts
@@ -3,6 +3,7 @@ import { DocumentNode } from "graphql";
 
 import { pageSearchFragment } from "../pageSearch/usePageSearch";
 import { pageTreePageFragment } from "../pageTree/usePageTree";
+
 export { GQLPageTreePageFragment } from "../pageTree/usePageTree";
 export { GQLPagesQuery, GQLPagesQueryVariables } from "./createPagesQuery.generated";