+++ packages/api/blocks-api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @comet/blocks-api
 
+## 5.2.0
+
+## 5.1.0
+
 ## 5.0.0
 
 ### Major Changes
@@ -10,7 +14,6 @@
 
     -   Install `@comet/cli` as a dev dependency
     -   Replace the scripts in the package.json of your admin:
-        
         ```json
             "generate-block-types": "comet generate-block-types --inputs",
             "generate-block-types:watch": "chokidar -s \"**/block-meta.json\" -c \"npm run generate-block-types\""